+++ packages/base/src/components/Tooltip/TooltipContent.tsx
@@ -6,7 +6,7 @@ import { Text } from '../Text';
 export const TooltipContent = ({ content, ...args }: TooltipContentProps) => {
   return (
     <StyledTooltipContent {...args}>
-      <Text typography='caption_400' color='neutral0' wordBreak='break-all'>
+      <Text typography='caption_400' color='neutral0' wordBreak='break-all' whiteSpace='pre-line'>
         {content}
       </Text>
     </StyledTooltipContent>